refactor player file organization
