+++ src/components/layout/index.ts
@@ -1,5 +1,3 @@
 export * from './delay/Delay';
 export * from './error-boundary/ErrorBoundary';
-export * from './PlayerBar/ExpandablePlayerBar';
 export * from './TabBar/TabBar';
-
