+++ src/features/player/hooks/index.ts
@@ -1,5 +1,5 @@
+export * from './usePlayer';
 export * from './usePlayerControls';
 export * from './usePlayerInitialization';
 export * from './usePlayerState';
 export * from './useTrackQueries';
-
