+++ src/features/player/views/components/progress-slider/hooks/useProgress.ts
