+++ src/features/player/views/hooks/usePlayerBarAnimation.ts
